BFS TRAVERSAL
The Breadth First Search (BFS) algorithm is used to search a graph data structure for a node that meets a set of criteria. 
It starts at the root of the graph and visits all nodes at the current depth level before moving on to the nodes at the next depth level.